Add loader file.
This patch moves code specific to loading egl
libraries and bindings from surface_factory_wayland
to a separate file. This will be extended to support
wayland specific bindings like EGL_WL_bind_wayland_display
and eglUnbindWaylandDisplayWL.
