+++ impl/surface_factory_wayland.cc
@@ -16,62 +16,7 @@
 #include "ozone/wayland/display.h"
 #include "ozone/wayland/window.h"
 #include "ozone/wayland/screen.h"
-#include "ui/gl/gl_bindings.h"
-#include "ui/gl/gl_egl_api_implementation.h"
-#include "ui/gl/gl_gl_api_implementation.h"
-#include "ui/gl/gl_implementation.h"
-#include "ui/gl/gl_switches.h"
-
-namespace gfx {
-namespace {
-
-// Load a library, printing an error message on failure.
-base::NativeLibrary LoadLibrary(const base::FilePath& filename) {
-  std::string error;
-  base::NativeLibrary library = base::LoadNativeLibrary(filename,
-                                                        &error);
-  if (!library) {
-    DVLOG(1) << "Failed to load " << filename.MaybeAsASCII() << ": " << error;
-    return NULL;
-  }
-  return library;
-}
-
-base::NativeLibrary LoadLibrary(const char* filename) {
-  return LoadLibrary(base::FilePath(filename));
-}
-
-}  // namespace
-
-bool InitializeGLBindings() {
-  base::NativeLibrary gles_library = LoadLibrary("libGLESv2.so.2");
-  if (!gles_library)
-    return false;
-  base::NativeLibrary egl_library = LoadLibrary("libEGL.so.1");
-  if (!egl_library) {
-    base::UnloadNativeLibrary(gles_library);
-    return false;
-  }
-
-  GLGetProcAddressProc get_proc_address =
-      reinterpret_cast<GLGetProcAddressProc>(
-          base::GetFunctionPointerFromNativeLibrary(
-              egl_library, "eglGetProcAddress"));
-  if (!get_proc_address) {
-    LOG(ERROR) << "eglGetProcAddress not found.";
-    base::UnloadNativeLibrary(egl_library);
-    base::UnloadNativeLibrary(gles_library);
-    return false;
-  }
-
-  SetGLGetProcAddressProc(get_proc_address);
-  AddGLNativeLibrary(egl_library);
-  AddGLNativeLibrary(gles_library);
-
-  return true;
-}
-
-}  // namespace gfx
+#include "ozone/wayland/egl/loader.h"
 
 namespace ui {
