+++ wayland/egl/loader.h
@@ -1,0 +1,10 @@
+// Copyright 2013 Intel Corporation. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+
+namespace gfx {
+
+bool InitializeGLBindings();
+
+}  // namespace gfx
